fix php 7.3 errors and warnings

+++ include/style.cls.php
@@ -466,7 +466,7 @@ class Style {
       }
         
       if ( $l === "normal" ) {
-        $ret += $ref_size;
+        $ret += (float) $ref_size;
         continue;
       }
 
@@ -498,7 +498,7 @@ class Style {
       }
       
       if ( ($i = mb_strpos($l, "%"))  !== false ) {
-        $ret += (float)mb_substr($l, 0, $i)/100 * $ref_size;
+        $ret += (float)mb_substr($l, 0, $i)/100 * (float) $ref_size;
         continue;
       }
 
@@ -524,7 +524,7 @@ class Style {
       
       // FIXME: em:ex ratio?
       if ( ($i = mb_strpos($l, "ex"))  !== false ) {
-        $ret += mb_substr($l, 0, $i) * $this->__get("font_size") / 2;
+        $ret += (float)mb_substr($l, 0, $i) * $this->__get("font_size") / 2;
         continue;
       }
       
@@ -1272,10 +1272,10 @@ class Style {
       return $this->_computed_border_radius;
     }
     
-    $rTL = $this->__get("border_top_left_radius");
-    $rTR = $this->__get("border_top_right_radius");
-    $rBL = $this->__get("border_bottom_left_radius");
-    $rBR = $this->__get("border_bottom_right_radius");
+    $rTL = (float)$this->__get("border_top_left_radius");
+    $rTR = (float)$this->__get("border_top_right_radius");
+    $rBL = (float)$this->__get("border_bottom_left_radius");
+    $rBR = (float)$this->__get("border_bottom_right_radius");
     
     if ( $rTL + $rTR + $rBL + $rBR == 0 ) {
       return $this->_computed_border_radius = array(
@@ -1287,10 +1287,10 @@ class Style {
       );
     }
     
-    $t = $this->__get("border_top_width");
-    $r = $this->__get("border_right_width");
-    $b = $this->__get("border_bottom_width");
-    $l = $this->__get("border_left_width");
+    $t = (float)$this->__get("border_top_width");
+    $r = (float)$this->__get("border_right_width");
+    $b = (float)$this->__get("border_bottom_width");
+    $l = (float)$this->__get("border_left_width");
     
     $rTL = min($rTL, $h - $rBL - $t/2 - $b/2, $w - $rTR - $l/2 - $r/2);
     $rTR = min($rTR, $h - $rBR - $t/2 - $b/2, $w - $rTL - $l/2 - $r/2);
